event/backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
pymysql==1.1.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
pydantic==2.5.0
qrcode[pil]==7.4.2
pillow==10.1.0
aiosmtplib==3.0.1
python-dotenv==1.0.0
email-validator==2.1.0